pyproject.toml
[tool.poetry]
name = "water"
version = "0.1.0"
description = ""
authors = ["nessnos <nesssnos@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.8"
django = "^5.1.1"
django-cors-headers = "^4.4.0"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
celery = "^5.4.0"
scikit-learn = "^1.6.1"
statsmodels = "^0.14.4"
redis = "^5.2.1"
langchain-community = "^0.3.19"
langchain-ollama = "^0.2.3"
tiktoken = "^0.9.0"
transformers = "^4.49.0"
langchain = "^0.3.20"
langchain-experimental = "^0.3.4"
sqlalchemy = "^2.0.38"
openai = "^1.65.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
